jobs/data-sync/pyproject.toml
[tool.poetry]
name = "strr-data-sync"
version = "1.0.0"
description = ""
authors = ["thorwolpert <thor@wolpert.ca>"]
packages = [{include = "strr_data_sync", from = "src"}]
license = "BSD 3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
flask = "^3.0.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
launchdarkly-server-sdk = "^9.0.1"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.44.1"
flask-babel = "^4.0.0"
psycopg2 = "^2.9.5"
flask-jwt-oidc = "^0.7.0"
gunicorn = "^21.2.0"
nanoid = "^2.0.0"
Jinja2 = "^3.1.2"
strr-api = {path = "../../strr-api", develop=true}
pandas = "^2.0.0"

[tool.poetry.group.test.dependencies]
freezegun = "^1.4.0"
hypothesis = "^6.100.0"
pexpect = "^4.9.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^23.12.1"
faker = "^24.4.0"
lovely-pytest-docker = "^0.3.1"
pylint = "^3.0.3"
bandit = "^1.7.6"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
skip_glob = ["devops"]
known_first_party = ["strr_data_sync", "strr_api"]


[tool.pylint.main]
fail-under = 10
max-line-length = 120
ignore = [ "migrations", "devops", "tests"]
ignore-patterns = ["^\\.#"]
ignored-modules= ["flask_sqlalchemy", "sqlalchemy", "SQLAlchemy" , "alembic", "scoped_session"]
ignored-classes= "scoped_session"
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"
extension-pkg-whitelist = "pydantic"
notes = ["FIXME","XXX","TODO"]
overgeneral-exceptions = ["builtins.BaseException", "builtins.Exception"]
confidence = ["HIGH", "CONTROL_FLOW", "INFERENCE", "INFERENCE_FAILURE", "UNDEFINED"]
disable = "C0209,C0301,W0511,W0613,W0703,W1514,R0801,R0902,R0903,R0911,R0401,R1705,R1718,W3101"
argument-naming-style = "snake_case"
attr-naming-style = "snake_case"
class-attribute-naming-style = "any"
class-const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
const-naming-style = "UPPER_CASE"
function-naming-style = "snake_case"
inlinevar-naming-style = "any"
method-naming-style = "snake_case"
module-naming-style = "any"
variable-naming-style = "snake_case"
docstring-min-length = -1
good-names = ["i", "j", "k", "ex", "Run", "_"]
bad-names = ["foo", "bar", "baz", "toto", "tutu", "tata"]
defining-attr-methods = ["__init__", "__new__", "setUp", "asyncSetUp", "__post_init__"]
exclude-protected = ["_asdict", "_fields", "_replace", "_source", "_make", "os._exit"]
valid-classmethod-first-arg = ["cls"]
valid-metaclass-classmethod-first-arg = ["mcs"]

[tool.flake8]
ignore = ["F401","E402", "Q000", "E203", "W503"]
exclude = [
    ".venv",
    ".git",
    ".history",
    "devops",
    "*migrations*",
    "tests",
]
per-file-ignores = [
    "__init__.py:F401",
    "*.py:B902"
]
max-line-length = 120
docstring-min-length=10
count = true

[tool.black]
target-version =  ["py310", "py311", "py312"]
line-length = 120
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  migrations
  | devops
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 